+++ components/cbaSettings/settings.sqf
@@ -5,7 +5,7 @@
 -------------------------------------------------------------------------------------------------------------------- */
 
 #include "..\..\optional_configuration\cbaSettings.hpp"
-#include "..\..\configuration\respawn.hpp"
+#include "..\..\core_configuration\respawn.hpp"
 
 //Medical
