+++ components/respawn/macros.hpp
@@ -1,5 +1,5 @@
 #include "..\..\macros.hpp"
 #include "..\..\downtime_macros.hpp"
-#include "..\..\configuration\respawn.hpp"
+#include "..\..\core_configuration\respawn.hpp"
 
 #define MINIMUM_RESPAWN_DELAY 5
