+++ core_configuration/endings.hpp
