+++ core_configuration/respawn.hpp
