+++ core_configuration/slottingGenerator.hpp
